chore: reorganize tests in the run package
