+++ packages/run/package.json
@@ -65,8 +65,8 @@
     "postpack": "rm -f oclif.manifest.json",
     "prepack": "rm -rf lib && tsc -b && oclif manifest",
     "pretest": "tsc -p test --noEmit",
-    "test": "nyc --extension .ts mocha --forbid-only \"test/**/*.test.ts\"",
-    "test:integration": "yarn test --grep='@acceptance'",
+    "test": "nyc --extension .ts mocha --forbid-only \"test/**/*.unit.test.ts\"",
+    "test:integration": "mocha --forbid-only \"test/**/*.integration.test.ts\"",
     "posttest": "yarn lint",
     "version": "oclif readme && git add README.md"
   }
