+++ packages/run/test/integration/logs.integration.test.ts
@@ -1,12 +1,6 @@
 import {expect, test} from '@oclif/test'
 
-let describeOrSkip: Mocha.SuiteFunction | Mocha.PendingSuiteFunction = describe
-
-if (process.env.CI && process.env.RUN_ACCEPTANCE_TESTS !== 'true') {
-  describeOrSkip = describe.skip.bind(describe)
-}
-
-describeOrSkip('@acceptance logs', () => {
+describe('logs', () => {
   test
     .stdout()
     .command(['logs', '--app=heroku-cli-ci-smoke-test-app'])
