+++ packages/run/test/unit/commands/run/detached.test.ts
