+++ packages/run/test/unit/lib/helpers.unit.test.ts
@@ -1,6 +1,6 @@
 import {expect, test} from '@oclif/test'
 
-import {buildCommand} from '../../src/lib/helpers'
+import {buildCommand} from '../../../src/lib/helpers'
 
 describe('helpers.buildCommand()', () => {
   [
@@ -11,8 +11,8 @@ describe('helpers.buildCommand()', () => {
     {args: ['echo', '{"foo":"bar"}'], expected: 'echo "{\\"foo\\":\\"bar\\"}"'},
   ].forEach(example => {
     test
-    .it(`parses \`${example.args.join(' ')}\` as ${example.expected}`, () => {
-      expect(buildCommand(example.args)).to.equal(example.expected)
-    })
+      .it(`parses \`${example.args.join(' ')}\` as ${example.expected}`, () => {
+        expect(buildCommand(example.args)).to.equal(example.expected)
+      })
   })
 })
